app/requirements.txt
flask
requests
bs4
selenium
chromedriver-binary-auto
mpu
flask-sqlalchemy
pymysql
flask-apscheduler
dotenv